+++ NEWS.md
@@ -11,6 +11,7 @@ Version 0.16.6b0 - dev
 - NEW: support for multi-line `ATTRIB` and `ATTDEF` entities in DXF R2018 
 - NEW: `Auditor` removes invalid DXF entities from layouts, blocks and the 
        OBJECTS section
+- NEW: `Auditor` removes standalone ATTRIB entities from layouts and blocks
 - CHANGE: renamed `RenderContext.current_layer` to `RenderContext.current_layer_properties` 
 - CHANGE: renamed `RenderContext.current_block_reference` to `RenderContext.current_block_reference_properties` 
 - CHANGE: extended entity validation for `GROUP`